Add `Margin` to `epaint`
Moved from `egui`
