+++ crates/egui/src/containers/frame.rs
@@ -1,6 +1,6 @@
 //! Frame container
 
-use crate::{layers::ShapeIdx, style::Margin, *};
+use crate::{layers::ShapeIdx, *};
 use epaint::*;
 
 /// Add a background, frame and/or margin to a rectangular background of a [`Ui`].
