+++ crates/epaint/src/lib.rs
@@ -27,6 +27,7 @@
 
 mod bezier;
 pub mod image;
+mod margin;
 mod mesh;
 pub mod mutex;
 mod shadow;
@@ -41,9 +42,10 @@ mod texture_handle;
 pub mod textures;
 pub mod util;
 
-pub use {
+pub use self::{
     bezier::{CubicBezierShape, QuadraticBezierShape},
     image::{ColorImage, FontImage, ImageData, ImageDelta},
+    margin::Margin,
     mesh::{Mesh, Mesh16, Vertex},
     shadow::Shadow,
     shape::{
